requirements.txt
aiotg==0.7.12
peewee-async==0.5.5
aiopg==0.11
requests==2.11.1
beautifulsoup4==4.5.1
pika==0.10.0